requirements.txt
asgiref
Django        
django-environ
django-filter 
djangorestframework
Markdown
pillow
sqlparse
tzdata